Add files via upload
A student management system (SMS) project, also known as a student information system (SIS) project, aims to manage and analyze information about a school's students. The system's purpose is to make it easier for parents, teachers, and administrators to track information and communicate with each other. SMS projects can help schools avoid disorganization, improve scheduling and activity clarity, and make it easier for school districts to use data for decision-making.